requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
python-decouple>=3.8

# MongoDB Integration
pymongo>=4.6.0
# djongo>=1.3.0  # Django 4.2 ile uyumlu değil

# Milvus Vector Database
pymilvus>=2.3.0

# Google Cloud Vision API
google-cloud-vision>=3.4.0
google-cloud-storage>=2.10.0

# Firebase Cloud Messaging
firebase-admin>=6.4.0
pyrebase4>=4.7.0

# Map APIs
googlemaps>=4.10.0
requests>=2.31.0

# Image Processing
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Additional utilities
celery>=5.3.0
redis>=5.0.0

# WebSocket Support
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Django Extensions
django-extensions>=3.2.0

# Additional Django packages
django-filter>=23.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7

# Database drivers
psycopg2-binary>=2.9.0

# Development tools
django-debug-toolbar>=4.2.0